+++ typescript/ui/src/plugins/SharePlugin/index.tsx
@@ -1,10 +1,14 @@
 import { lazy } from "react";
 import { createPlugin } from "../../core/utils/plugins/utils";
+import { makeSlot } from "../../core/utils/slots/utils";
+
+const ShareButton = lazy(() => import("./components/ShareButton"));
 
 export const SharePluginName = "SharePluginName";
 export const SharePlugin = createPlugin({
   name: SharePluginName,
   components: {
-    ShareButton: lazy(() => import("./components/ShareButton")),
+    ShareButton,
   },
+  slots: [makeSlot("layout.headerActions", ShareButton, 5)],
 });